backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
pydantic>=2.6.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
celery>=5.3.6
redis>=5.0.1
httpx>=0.26.0
alembic>=1.13.1
google-generativeai>=0.3.0
